server/requirements.txt
alembic
fastapi[all]
passlib[bcrypt]
psycopg2
pyjwt
redis
sqlalchemy
pyotp
qrcode